Add first unit test
Add the `register_user` method on `FilmBot` that will attempt to
register a Discord user in the DataBase.  Use `moto` to mock out `boto3`
and unit test this function.  Add a GitHub workflow to run unit tests on
all pull requests.

An initial sketch on the DynamoDB schema at the top of `filmbot.py`.

Fix incorrect Python dependency `nacl` to `pynacl`.