Implement start_watching_film and schema change
In order to implement attendance voting we need to be able to sort
our films by the date that they were watched.  This way we can query for
the last watched film and see whether we fall within its watch range.

To do this we change the database schema to have the Discord Guild ID be
the partition key and have our film ID/user ID be the sort key.  Based
on AWS recommendations, I have combined our tables into one and can be
differentiated with the sort key.