Refactor into classes
Stop passing around random dictionaries and pass `User` and `Firm`
instead.  Move out some of our query code to a `__query` method to
reduce some code.

This commit also updates some of the Discord messages and updates
`start_watching_film` to return the name of the film.  These should
probably be as a separate commit, but what's done is done.