Fix nomination message
The message incorrect states that the person who ran `/watch` should be
the person to nominate the next film.  Instead it should get the
`DiscordUserID` for that film and use their name.

To do this we return an entire `Film` object from `start_watching_film`
and grab the nominator from that.