Add IMDb integration
Add IMDb-powered autocomplete for the `/nominate` application command
using `PyIMDb`.  If the user doesn't use the autocomplete, then they
will add a film without a corresponding IMDb movie ID.

When handling the `/nominate` command, we can't call `ia.get_movie` and
get the title without exceeding the 3s Discord response window.  To
avoid changing the architecture too much, we encode both the film namd
and IMDb movie ID in the autocomplete value and trust that this is
accurrate.  It is possible for a user to manually type a film in the
format `IMDB:[movie id]:[film name]` with mismatching movie ID and film
name.  However, we are working with (somewhat) trusted users so this
isn't too much of a risk.

Store the IMDb movie ID in the database and display a clickable link
when listing out the current nominations.