Optimize hot starts with static initialization
If we store the `boto3` client as a global then it will get created once
when the AWS Lambda cold starts and then can be reused for all further
hot starts.  Time spent in this initialization phase is also not counted
for AWS Lambda cost.

Source: https://aws.amazon.com/blogs/compute/operating-lambda-performance-optimization-part-2/