Add /history application command
Add a way to list all of the films that have been watched in reverse
chronological order.

We are limited to 2000 characters in a Discord message so at some point
we won't show all of the films with `/history`.  In that case we should
most likely add a "View More" button at the end of the message that
displays the next chunk of films.