Add a public /naughty option
Add a button to the output of `/naughty` that posts the same message but
to all users.  The button only appears if there are outstanding tasks.

Right now it's possible to spam this button, which isn't ideal.  When
pressing the button we should edit the original message in Discord if we
are still within the 15 minute time frame and disable it.  Otherwise we
should also not publically shame if more than 15 minutes have passed.
In this case we would restrict a public shaming to once per `/naughty`
command.  For the moment we have to trust users not to abuse it too much
:)