Restrict `moto` version
`moto` 3 fixed an issue where `ScanIndexForward` and filters were
applied in the wrong order (https://github.com/getmoto/moto/issues/3909)
and causes incorrect results when we try to query for the latest film.

Both `moto` version 4 or 5 seem to have the same issue.

Update the `mock_dynamodb2` import with `mock_dynamodb` as they are
identical and the latter is scheduled for removal.