Change user sort key and improve `/peek`
Improve `/peek` by displaying users who have not nominated any films
yet.  This is part of the way to moving all the information in
`/naughty` to `/peek`.

In order to support this, we need to be able to query the film name and
vote count by user.  We probably should be duplicating this information
into the user record, which is the noSQL way.  However, if we change the
sort key for users from `USER#` to `DISCORDUSER#`, then we are able to
grab all users and `FILM#NOMINATION#` entries with one query because
they would be next to each other alphabetically.

In an unrelated change - that really should be a separate change - we
fix some of the `datetime` imports that no longer work on Python 3.9.