Fix `/history` being cut off
Due to the 2Kb limit in Discord messages we are now at the point where
the watching history returned in `/history` is incomplete.

To improve the number of entries returned, we remove the IMDb link,
which takes up about 50% of the message limit.

If there are additional items, display a button message component that
will display the next set of entries.